+++ lib/modules/i18n/i18n_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: i18n
-// Updated at: 2026-05-14 23:14:27
-// Commit: chore(i18n): update dependencies and cleanup unused imports in i18n
+// Updated at: 2026-05-15 17:00:57
+// Commit: test(i18n): add unit tests for i18n business logic
 
 class i18nComponentState {
-  final String commitRef = '1171';
-  final String lastMessage = 'chore(i18n): update dependencies and cleanup unused imports in i18n';
-  final DateTime timestamp = DateTime.parse('2026-05-14T23:14:27+05:00');
+  final String commitRef = '1188';
+  final String lastMessage = 'test(i18n): add unit tests for i18n business logic';
+  final DateTime timestamp = DateTime.parse('2026-05-15T17:00:57+05:00');
 }